Revert "feat: remat"

+++ omegalax/models/qwen3_vl/config.py
@@ -24,7 +24,7 @@ class Qwen3VLVisionConfig:
     hidden_act: str
     num_position_embeddings: int
     deepstack_visual_indexes: tuple[int, ...]
-    dtype: Any = jnp.bfloat16
+    dtype: Any = jnp.float32
 
 
 @dataclasses.dataclass(frozen=True, slots=True)
